# 3% faster radix sort