@LilithHafner 2023-02-03
Thanks! A 3% performance increase is negligible with respect to noise and systemic error in most cases but this arithmetic refactoring is correct, adds no code complexity (arguably simplifies the code by making the code in the hot loop simpler) and is clearly a performance improvement, even if small.